% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Str_C_Operator.R
\name{\%+\%}
\alias{\%+\%}
\alias{\%+\%}
\title{Operator for concatenation}
\usage{
\%+\%(a, b)
}
\arguments{
\item{a}{The first String}

\item{b}{The second String}
}
\description{
Wrapper for stringr::str_c()
}
\details{
Wrapper for stringr::str_c() used for quicker pasting of Strings
}
