% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callRevFromTerminal.R
\name{callRevFromTerminal}
\alias{callRevFromTerminal}
\title{Execute a .rev file in RevBayes through an RStudio terminal}
\usage{
callRevFromTerminal(revscript)
}
\arguments{
\item{revscript}{.rev file to execute in RevBayes}
}
\value{
termID Unique identifier of the terminal used to call RevBayes
}
\description{
Accesses an RStudio terminal through the RStudio API and runs a .rev script.
This allows users to submit mcmcs and longer scripts after writing them in R Markdown through knitr or interactively with repRev()
}
\examples{
\dontrun{
saveRev("archertutorial.rev", use_quit=TRUE)
callRevFromTerminal("archertutorial.rev")
}

}
