% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCallRev.R
\name{cleanCallRev}
\alias{cleanCallRev}
\title{Wrapper for callRev(). Runs previous code in revEnv$allCode to allow the user to
refer to objects that have been defined in rb but not in the revEnv.}
\usage{
cleanCallRev(input)
}
\arguments{
\item{input}{Code snippet to be ran in rb.exe.}
}
\description{
Wrapper for callRev(). Runs previous code in revEnv$allCode to allow the user to
refer to objects that have been defined in rb but not in the revEnv.
}
