% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearRev.R
\name{clearRev}
\alias{clearRev}
\title{Removes lines from RevBayes history}
\usage{
clearRev(n = NULL)
}
\arguments{
\item{n}{How many lines to remove. If n = NULL, all lines are removed.}
}
\description{
Removes lines of code from the .Revhistory file used for managing RevBayes interactions.
}
\examples{
\dontrun{
clearRev() #Clear all objects from RevBayes
clearRev(n = 1) # Clear the last line input to RevBayes
}
}
