% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectRevObjects.R
\name{detectRevObjects}
\alias{detectRevObjects}
\title{Locates names of objects from revenv located in a string. Object names must be surrounded by '#'.}
\usage{
detectRevObjects(string)
}
\arguments{
\item{string}{string vector where rev object names will be located}
}
\value{
revObjs: Names of objects from revenv found in string
}
\description{
Locates names of objects from revenv located in a string. Object names must be surrounded by '#'.
}
