% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRev.R
\name{doRev}
\alias{doRev}
\title{Wrapper for cleanCallRev(). Makes running multiple lines of rb code cleaner by splitting input by ";".}
\usage{
doRev(input, viewCode = FALSE)
}
\arguments{
\item{input}{Code snippet to be ran in rb.exe.}

\item{viewCode}{see Rev code input and output in the viewing pane}
}
\description{
Wrapper for cleanCallRev(). Makes running multiple lines of rb code cleaner by splitting input by ";".
}
