% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRev.R
\name{findRev}
\alias{findRev}
\title{Searches for the RevBayes executable}
\usage{
findRev(parentDirectory = "~")
}
\arguments{
\item{parentDirectory}{Parent directory to recursively search for the rb executable. Default is the user's root directory ('~').}
}
\description{
Recursively earches the users machine for the RevBayes executable and returns a vector of possible paths
}
\examples{
\dontrun{
findRev()
}

}
