% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMS.R
\name{getMS}
\alias{getMS}
\title{coerces R list or matrix objects from revenv into a string format that can be read in rb}
\usage{
getMS(outMatrix)
}
\arguments{
\item{outMatrix}{name of list or matrix object in revenv surrounded by '#' and '#'}
}
\value{
coercedMatrix: String format of outMatrix that can be read into rb
}
\description{
coerces R list or matrix objects from revenv into a string format that can be read in rb
}
\examples{
getMS("#myMatrix#")

}
