% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRev.R
\name{getRev}
\alias{getRev}
\title{Get variable definitions from RevEnv}
\usage{
getRev()
}
\description{
Prints revEnv$Vars to return definitions of variables that have been defined in rb.exe
   using a Revticulate function.
}
