% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevHistory.R
\name{getRevHistory}
\alias{getRevHistory}
\title{Get Revticulate History}
\usage{
getRevHistory()
}
\value{
lines - character. A vector of the line-by-line Revticulate history
}
\description{
Return the history of RevBayes code submitted with doRev() and similar functions
}
\examples{
\dontrun{
getRevHistory()

cat(getRevHistory(), sep="\n")
}

}
