% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevObj.R
\name{getRevObj}
\alias{getRevObj}
\title{Getter function to retrieve objects created in rb.exe.}
\usage{
getRevObj(name, coerce = FALSE)
}
\arguments{
\item{name}{String. Name of object to retrieve}

\item{coerce}{boolean. If true, Revticulate attempts to coerce the String formatted Rev object into
a comparable R object. Default is FALSE.}
}
\description{
Getter function to retrieve objects created in rb.exe.
}
