% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevVars.R
\name{getRevVars}
\alias{getRevVars}
\title{Show variables that have been defined in RevBayes}
\usage{
getRevVars()
}
\value{
No return - wrapper for RevBayes ls() function to show variables in rb environment
}
\description{
Wrapper for RevBayes ls() function. Displays variables that have been defined in the working RevBayes session.
}
\examples{
\dontrun{
getRevVars()
}

}
