% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSS.R
\name{getSS}
\alias{getSS}
\title{coerces R string object from revenv into a string format that can be read in rb}
\usage{
getSS(outString)
}
\arguments{
\item{outString}{name of string object in revenv surrounded by '#' and '#'}
}
\value{
coerceString: String format of outString that can be read into rb
}
\description{
coerces R string object from revenv into a string format that can be read in rb
}
\examples{
getSS("#myString#")

}
