% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTS.R
\name{getTS}
\alias{getTS}
\title{coerces R phylo objects from revenv into a string format that can be read in rb}
\usage{
getTS(outTree)
}
\arguments{
\item{outTree}{name of phylo object in revenv surrounded by '#' and '#'}
}
\value{
revSTR: String format of outTree that can be read into rb
}
\description{
coerces R phylo objects from revenv into a string format that can be read in rb
}
\examples{
getTS("#myTree#")

}
