% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVS.R
\name{getVS}
\alias{getVS}
\title{coerces R vector object from revenv into a string format that can be read in rb}
\usage{
getVS(outVector)
}
\arguments{
\item{outVector}{name of vector object in revenv surrounded by '#' and '#'}
}
\value{
revVec: String format of outVector that can be read into rb
}
\description{
coerces R vector object from revenv into a string format that can be read in rb
}
\examples{
getVS("#myVector#")

}
