% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Str_C_Operator.R
\name{\%+\%}
\alias{\%+\%}
\title{Operator to streamline pasting. Uses stringr::str_c for speed}
\usage{
a \%+\% b
}
\arguments{
\item{a}{The first string to concatenate}

\item{b}{The second string to concatenate}
}
\description{
Operator to streamline pasting. Uses stringr::str_c for speed
}
