% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitRev.R
\name{initRev}
\alias{initRev}
\title{Initializes external variables for interaction with RevBayes}
\usage{
initRev(searchPath = "~")
}
\arguments{
\item{searchPath}{Full path or directory to search for the RevBayes executable. Default is the user's root directory (~).}
}
\description{
Creates external variables for storing the paths to the RevBayes executable and .Revhistory files, as well as initiating a folder to store temp files for RevBayes interactions.
}
\examples{
\dontrun{
RevPath <- "C://Users/Caleb/Documents/WrightLab/RevBayes_Win_v1.0.13/RevBayes_Win_v1.0.13/rb.exe"
initRev(RevPath)
}
}
