% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitRev.R
\name{KnitRev}
\alias{KnitRev}
\title{knitr engine for revbayes}
\usage{
KnitRev()
}
\arguments{
\item{options}{argument required for knitr engines. options$code coerces the text
in the code chunk to a string, which is then evaluated with doRev()}
}
\description{
knitr engine for revbayes
}
