% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRevHistory.R
\name{loadRevHistory}
\alias{loadRevHistory}
\title{Load in Rev history from a .Rev file}
\usage{
loadRevHistory(filepath, overwrite = FALSE)
}
\arguments{
\item{filepath}{File to load .Revhistory from}

\item{overwrite}{If TRUE, history load from filepath will overwrite current .Revhistory. If FALSE, it will be appended to the end of it.}
}
\description{
Load in Rev history from a .Rev file
}
\examples{
\dontrun{
}


}
