% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRevObjects.R
\name{repRevObjects}
\alias{repRevObjects}
\title{Locates objects from revenv in a string and replaces that part of the string with a string reperesentation of the object that can be read in rb}
\usage{
repRevObjects(string)
}
\arguments{
\item{string}{String containing names of revenv objects surrounded by '#' and '#'}
}
\value{
The origional string coerced into a format that will read the revenv objects into rb
}
\description{
Locates objects from revenv in a string and replaces that part of the string with a string reperesentation of the object that can be read in rb
}
