% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevDefine.R
\name{RevDefine}
\alias{RevDefine}
\title{Defines rev objects and returns them to revenv.}
\usage{
RevDefine(RevOut, viewCode = FALSE)
}
\arguments{
\item{RevOut}{Rev code string defining object}

\item{viewCode}{If TRUE, rev code and output will be shown in stringr::str_view widget}
}
\description{
Defines rev objects and returns them to revenv.
}
\examples{
RevDefine('var1 <- posteriorPredictiveProbability(v(2), 3)', viewCode = TRUE)

}
