% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Str_C_Operator.R
\name{\%+\%}
\alias{\%+\%}
\title{Operator for concatenation.}
\usage{
\%+\%(a, b)
}
\arguments{
\item{a}{The first String}

\item{b}{The second String}
}
\description{
Rapid string parsing
}
\details{
Method for quicker pasting of Strings.
}
