% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undoRev.R
\name{undoRev}
\alias{undoRev}
\title{Undo n number of previous Rev commands.}
\usage{
undoRev(n)
}
\arguments{
\item{n}{Number of prior unputs to remove from allCode.}
}
\description{
Undo the last specified number of unputs from allCode. Prints revEnv$allCode to
the screen to show the user the new Rev code history.
}
