% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallRev.R
\name{CallRev}
\alias{CallRev}
\title{Calls function from rb and returns object in R}
\usage{
CallRev(..., coerce = TRUE, path = revenv$RevPath, viewCode = F)
}
\arguments{
\item{...}{Command to be passed into rb}

\item{coerce}{If TRUE, attempts to coerce output into R object. if FALSE, returns raw string output from rb.}

\item{path}{Path to rb. If InitRev has been called, defaults to revenv.RevPath.}

\item{viewCode}{If TRUE, uses stringr::str_view_all function to view raw input and output sent to rb}
}
\value{
out: Output object from rb. If coerce = TRUE, out will be an R equilvalent of rev object. If FALSE, out will be string representation of the object.
}
\description{
Calls function from rb and returns object in R
}
