% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallRev.R
\name{callRev}
\alias{callRev}
\title{Submit input to rb.exe and return output}
\usage{
callRev(
  ...,
  coerce = FALSE,
  path = Sys.getenv("RevBayesPath"),
  viewCode = F,
  use_wd = T,
  knit = F,
  timeout = 5
)
}
\arguments{
\item{...}{String input to send to RevBayes.}

\item{coerce}{If TRUE, attempts to coerce output to an R object. If FALSE, output
will remain in String format. Default is FALSE.}

\item{path}{Path to rb.exe. Default is Sys.getEnv("RevBayesPath"), which is created with initRev().}

\item{viewCode}{If TRUE, the input input and output in the temporary file used to interact
with rb.exe will be displayed in the viewing pane. This option may be useful for
diagnosing errors.}

\item{use_wd}{If TRUE, sets the working directory in the temporary RevBayes session
to the working directory of the active R session.}

\item{knit}{Argument used to manage output formatting for knitRev(). This argument
should generally be ignored by the user.}

\item{timeout}{Determines how long the system2() call should wait before timing out (seconds). Default is 5.}
}
\value{
out Output from RevBayes. If coerce = FALSE, out will be in String format.
If coerce = TRUE, the function will attempt to coerce the String to an R object.
}
\description{
Submits input to rb.exe and returns output to R in string format. If coerce = T, the function
will try to coerce output to a similar R object.
}
\examples{
\dontrun{
callRev("2^3")
callRev("2^3", coerce = FALSE, viewcode = T)
}
}
