% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallRev.R
\name{CallRev}
\alias{CallRev}
\title{Submit input to rb.exe and return output}
\usage{
CallRev(..., coerce = TRUE, path = RevEnv$RevPath, viewCode = F, use_wd = T)
}
\arguments{
\item{...}{String input to send to RevBayes.}

\item{coerce}{If TRUE, attempts to coerce output to an R object. If FALSE, output
will remain in String format. Default is TRUE.}

\item{path}{Path to rb.exe. Default is RevEnv$RevPath, which is created with InitRev().}

\item{viewCode}{If TRUE, the input input and output in the temporary file used to interact
with rb.exe will be displayed in the viewing pane. This option may be useful for
diagnosing errors.}

\item{use_wd}{If TRUE, sets the working directory in the temporary RevBayes session
to the working directory of the active R session.}
}
\value{
out Output from RevBayes. If coerce = FALSE, out will be in String format.
    If coerce = TRUE, the function will attempt to coerce the String to an R object.
}
\description{
Submits input to rb.exe and returns output to R in string format. If coerce = T, the function
   will try to coerce output to a similar R object.
}
\examples{
CallRev("2^3")
CallRev("2^3", coerce = FALSE, viewcode = T)

}
