% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callRevFromTerminal.R
\name{callRevFromTerminal}
\alias{callRevFromTerminal}
\title{Execute a .rev file in RevBayes through the R studio terminal}
\usage{
callRevFromTerminal(revscript)
}
\arguments{
\item{revscript}{.rev file to execute in RevBayes}
}
\value{
termID Unique identifier of the terminal used to call RevBayes
}
\description{
Execute a .rev file in RevBayes through the R studio terminal
}
