% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callRevFromTerminal.R
\name{callRevFromTerminal}
\alias{callRevFromTerminal}
\title{Execute a .rev file in RevBayes Through an RStudio Terminal}
\usage{
callRevFromTerminal(revscript)
}
\arguments{
\item{revscript}{.rev file to execute in RevBayes}
}
\value{
termID Unique identifier of the terminal used to call RevBayes
}
\description{
Accesses an RStudio terminal through the RStudio API and runs a .rev script.
}
\details{
Many common RevBayes use-cases, such as generating mcmcs, take a long time to execute.
Because of this fact, it is inefficient and inadvisable to run them in knitr documents
or with RevBayes calls through repRev(). Instead, tutorial code or code that has been written
interactively can be externally saved with saveRev(), and then executed in an RStudio terminal
via callRevFromTerminal(). Output files can then be explored and visualized with the
RevGadgets package.
}
\examples{
\dontrun{
saveRev("archertutorial.rev", use_quit=TRUE)
callRevFromTerminal("archertutorial.rev")
}

}
