% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearRev.R
\name{clearRev}
\alias{clearRev}
\title{Empties revEnv}
\usage{
clearRev(n = NULL)
}
\arguments{
\item{n}{How many lines to remove. Should be an integer}
}
\description{
Clears all code, or a user-specified amount of code, and vars from revEnv.
}
\examples{
\dontrun{
clearRev() #Clear all objects from RevBayes
clearRev(n = 1) # Clear the last line input to RevBayes
}
}
