% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceRev.R
\name{CoerceRev}
\alias{CoerceRev}
\title{Coerces string of RevBayes output into an equivalent R object.}
\usage{
CoerceRev(out)
}
\arguments{
\item{...}{out}
}
\value{
out Coerce output of the RevBayes output string.
}
\description{
Coerces string of RevBayes output into an equivalent R object.
Uses the structure of the characters within the string to identify an appropiate R object type (a list, vector,
or even phylo tree) to coerce the string into.
}
\examples{
CoerceRev("[1, 2, 3, 4]")

}
