% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceRev.R
\name{coerceRev}
\alias{coerceRev}
\title{Convert RevBayes Output into R Objects}
\usage{
coerceRev(revString)
}
\arguments{
\item{revString}{character - Output from RevBayes}
}
\value{
rObject: Type varies depending on Rev object type. R object-formatted output coerced from a RevBayes output string.
}
\description{
Coerces a RevBayes output string into an equivalent R object.
Automatically determines R object type based on the structure of the string,
and returns the original string if R object equivalent cannot be determined.
}
\examples{
\dontrun{
coerceRev("[1, 2, 3, 4]")
}
}
