% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoerceRev.R
\name{coerceRev}
\alias{coerceRev}
\title{Coerces string of RevBayes output into an equivalent R object.}
\usage{
coerceRev(out)
}
\arguments{
\item{out}{String formatted representation of a Rev language object.
coerceRev() will recognize the formatting of most commonly used
Rev objects and will convert them into equivalent R objects. If coerceReV()
does not recognize the objects formatting, the initial String representation
will be returned.}
}
\value{
out Coerce output of the RevBayes output string.
}
\description{
Coerces string of RevBayes output into an equivalent R object.
Uses the structure of the characters within the string to identify an appropiate R object type (a list, vector,
or even phylo tree) to coerce the string into.
}
\examples{
\dontrun{
coerceRev("[1, 2, 3, 4]")
}
}
