% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRev.R
\name{doRev}
\alias{doRev}
\title{Wrapper for callRev(). Runs previous code in the .Revhistory file to allow user-created Rev variables to persist between interactions.}
\usage{
doRev(input, viewCode = FALSE, coerce = FALSE, timeout = 5)
}
\arguments{
\item{input}{Code snippet to run in the RevBayes executable}

\item{viewCode}{If true, Rev code input and output will be displayed in the viewing pane.}

\item{coerce}{If true, the output from RevBayes will be coerced into R format with coerceRev()}

\item{timeout}{Determines how long the system2() call should wait before timing out (seconds). Default is 5.}
}
\description{
Wrapper for callRev(). Runs previous code in the .Revhistory file to allow user-created Rev variables to persist between interactions.
}
