% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRev.R
\name{doRev}
\alias{doRev}
\title{Wrapper for callRev(). Runs previous code in the .Revhistory file to allow user-created Rev variables to persist between interactions.}
\usage{
doRev(input, coerce = TRUE, evaluate = TRUE, viewCode = FALSE, timeout = 5)
}
\arguments{
\item{input}{Code snippet to run in the RevBayes executable}

\item{coerce}{If TRUE, the output from RevBayes will be coerced into R format with coerceRev()}

\item{viewCode}{If TRUE, Rev code input and output will be displayed in the viewing pane.}

\item{timeout}{Determines how long the system2() call should wait before timing out (seconds). Default is 5.}
}
\value{
now: type varies. If coerce = TRUE, coerceRev() will attempt to convert RevBayes output into an equivalent R object. Else, return type is character.
}
\description{
Wrapper for callRev(). Runs previous code in the .Revhistory file to allow user-created Rev variables to persist between interactions.
}
