% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRev.R
\name{doRev}
\alias{doRev}
\title{Wrapper for callRev(). Runs previous code in revEnv$allCode to allow the user to
refer to objects that have been defined in rb but not in the revEnv.}
\usage{
doRev(input, viewCode = FALSE, coerce = FALSE)
}
\arguments{
\item{input}{Code snippet to be ran in rb.exe.}

\item{viewCode}{see Rev code input and output in the viewing pane}
}
\description{
Wrapper for callRev(). Runs previous code in revEnv$allCode to allow the user to
refer to objects that have been defined in rb but not in the revEnv.
}
