% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRev.R
\name{doRev}
\alias{doRev}
\title{Wrapper for CallRev() and RevDefine()}
\usage{
doRev(
  ...,
  trackCode = FALSE,
  viewCode = FALSE,
  coerce = TRUE,
  interactive = FALSE,
  Det_Update = TRUE,
  use_wd = T,
  knit = FALSE
)
}
\arguments{
\item{...}{String of code that will be sent to rb.exe}

\item{trackCode}{If true, Rev code will be will be appended to RevEnv$allCode.}

\item{viewCode}{If TRUE, code from the temporary file used to interact with rb.exe
will be displayed in the viewing pane. This option may be useful for diagnosing errors.
Default is FALSE.}

\item{coerce}{If FALSE, output from rb.exe will remain in String format. If TRUE, doRev()
will attempt to coerce output into an appropriate R object.}

\item{interactive}{Ignore. Used to implement doRev() into RepRev().}

\item{Det_Update}{If TRUE, previously created object in RevEnv that is redefined in rb.exe
will be updated. If FALSE, it will not. Default is TRUE.}

\item{use_wd}{If TRUE, the temporary rb.exe session will use the same working directory as
the current R session. Else, it will use its default.}
}
\value{
outobjs Prints output from rb.exe, if required.
}
\description{
A wrapper for CallRev() and RevDefine(). If a variable is assigned in the temporary
   rb session, it will also be created in the RevEnv. Otherwise, functions like CallRev().
}
\examples{
doRev("sam <- 777")
doRev("sam + 2", viewCode = TRUE)

}
