% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRev.R
\name{findRev}
\alias{findRev}
\title{Searches for the RevBayes executable}
\usage{
findRev()
}
\description{
Searches the users machine for the RevBayes executable and returns a vector of possible paths
}
\examples{
\dontrun{
findRev()
}

}
