% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRev.R
\name{GetRev}
\alias{GetRev}
\title{Get variable definitions from RevEnv}
\usage{
GetRev()
}
\description{
Prints RevEnv$Vars to return definitions of variables that have been defined in rb.exe
   using Revticulate function
}
