% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevHistory.R
\name{getRevHistory}
\alias{getRevHistory}
\title{Get full history of RevBayes code from RevEnv}
\usage{
getRevHistory()
}
\value{
lines: character. Lines read from .Revhistory file.
}
\description{
Returns a vector of the lines of code in the .Revhistory file
}
