% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevObj.R
\name{getRevObj}
\alias{getRevObj}
\title{Getter function to retrieve objects created in RevBayes}
\usage{
getRevObj(name, coerce = FALSE)
}
\arguments{
\item{name}{String. Name of object to retrieve}

\item{coerce}{Boolean. If true, Revticulate attempts to coerce the String formatted Rev object into
a comparable R object. Default is FALSE.}
}
\value{
revObject: type varies. Object returned from RevBayes that was previously defined in .Revhistory. If coerce == FALSE, returns character. Else, returns type determined by coerceRev().
}
\description{
Getter function to retrieve objects created in RevBayes
}
