% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevVars.R
\name{getRevVars}
\alias{getRevVars}
\title{Get variable definitions from RevEnv}
\usage{
getRevVars(varName = NULL)
}
\arguments{
\item{varName}{Only returns lines where specific variable names are defined. If NULL, all lines are returned}
}
\description{
Returns vector of lines in .Revhistory containing variable definitions
}
\examples{
\dontrun{
getRevVars()
}
\dontrun{
getRevVars("var1")
}

}
