% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRevVars.R
\name{getRevVars}
\alias{getRevVars}
\title{Get variable definitions from RevEnv}
\usage{
getRevVars()
}
\description{
Prints revEnv$vars to return definitions of variables that have been defined in rb.exe
   using a Revticulate function.
}
