% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitRev.R
\name{InitRev}
\alias{InitRev}
\title{Create an environment for interacting with RevBayes}
\usage{
InitRev(path)
}
\arguments{
\item{path}{String path to rb.exe}
}
\description{
Create an environment for interacting with RevBayes
}
\examples{
RevPath <- "C://Users/Caleb/Documents/WrightLab/RevBayes_Win_v1.0.13/RevBayes_Win_v1.0.13/rb.exe"
InitRev(RevPath)

}
