% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitRev.R
\name{InitRev}
\alias{InitRev}
\title{Inititates an enironment to store Rev objects and sets path to rb}
\usage{
InitRev(path)
}
\arguments{
\item{path}{String path to rb.}
}
\description{
Inititates an enironment to store Rev objects and sets path to rb
}
