% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitRev.R
\name{initRev}
\alias{initRev}
\title{Create an environment for interacting with RevBayes}
\usage{
initRev(path = NULL, useHistory = FALSE)
}
\arguments{
\item{path}{String path to rb.exe}

\item{useHistory}{boolean Should the code from the previous Revticulate session be written into
revEnv$allCode? Default is FALSE.}
}
\description{
Creates an environment for interecting with rb.exe. This environment contains the variable
RevPath for storing the path to rb.exe. If no path is provided, it will default to trying
to use a path stored in Revticulate/RevPath.txt in .libPaths(). This means the user only has
to provide the path to rb.exe once, and can change the path at any time by applying it again
in initRev().
}
\examples{
\dontrun{
RevPath <- "C://Users/Caleb/Documents/WrightLab/RevBayes_Win_v1.0.13/RevBayes_Win_v1.0.13/rb.exe"
initRev(RevPath)
}
}
