% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitRev.R
\name{KnitRev}
\alias{KnitRev}
\title{Knitr engine for RevBayes}
\usage{
KnitRev()
}
\arguments{
\item{options}{Argument required for Knitr engines. options$code coerces the text
in the code chunk to a string, which is then evaluated with doRev().}
}
\description{
Knitr engine for RevBayes
}
