% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitRev.R
\name{knitRev}
\alias{knitRev}
\title{Knitr engine for RevBayes}
\usage{
knitRev()
}
\description{
Rev code can be ran directly in knitr chunks, without the use of the functions doRev() or repRev(). History is accessed with the .Revhistory file and persists between chunks.
}
