% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnitRev.R
\name{knitRev}
\alias{knitRev}
\title{Knitr engine for RevBayes}
\usage{
knitRev()
}
\description{
Rev code is ran directly in knitr chunks, and using the wrapper functions isn't necessary.
Any created variables will be put in RevEnv, and defined variables can be used across multiple
chunks.
}
