% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRevHistory.R
\name{loadRevHistory}
\alias{loadRevHistory}
\title{Load Revticulate History from a .Rev File}
\usage{
loadRevHistory(filepath, overwrite = FALSE)
}
\arguments{
\item{filepath}{character - File to load .Revhistory from}

\item{overwrite}{logical - If TRUE, the code in the file at 'filepath' will overwrite current .Revhistory. If FALSE, it will be appended to the end of it.}
}
\value{
No return - reads in new code to the .Revhistory file
}
\description{
Reads in the code from a .Rev file and writes it into the current Revticulate history.
}
\examples{
\dontrun{
     loadRevHistory("someCode.rev")

     loadRevHistory("someMoreCode.rev", overwrite=TRUE)
}

}
