% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepRev.R
\name{RepRev}
\alias{RepRev}
\title{Established an interactive rev session using CallRev. To break, type "quit()"}
\usage{
RepRev(viewCode = F)
}
\arguments{
\item{viewCode}{If TRUE, uses stringr::str_view_all function to view raw input and output sent to rb}
}
\description{
Established an interactive rev session using CallRev. To break, type "quit()"
}
