% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevDefine.R
\name{RevDefine}
\alias{RevDefine}
\title{Define Rev objects into RevEnv when calling rb.exe}
\usage{
RevDefine(RevOut, viewCode = FALSE, hideMessage = FALSE, knit = FALSE)
}
\arguments{
\item{RevOut}{String input that will be sent to rb.exe. If the function detects an
an assignment operator, a similar R object to the one being created in the rb session
will be created in the RevEnv.}

\item{viewCode}{If TRUE, the code written in the temporary file used to interact with
rb.exe will be displayed in the viewing pane. This may be useful for diagnosing errors.}

\item{hideMessage}{If FALSE, when RevDefine() creates an object in the RevEnv a message
will be displayed. If TRUE, the message will be hidden.}
}
\description{
A wrapper for CallRev() used when objects are defined in RevBayes. Identifies if
    an object is being defined, and if it is it will create an object of the same
    name in RevEnv. Because the object is being created in RevEnv, InitRev() must be called
    first.
}
\examples{
RevDefine("example1 <- posteriorPredictiveProbability(v(2), 3)", viewCode = TRUE)

}
