% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRev.R
\name{saveRev}
\alias{saveRev}
\title{Save Revticulate history in a .Rev file for later use}
\usage{
saveRev(filepath, use_wd = TRUE, use_quit = TRUE)
}
\arguments{
\item{filepath}{File to write the RevBayes history to}

\item{use_wd}{If TRUE, the history will be prepended by setwd('users current working directory'). This
is especially helpful when writing a file for running an mcmc, as any Moniters would otherwise
write output to the parent direcory of the RevBayes executable. Default is TRUE.}

\item{use_quit}{If TRUE, q() will be appended to the end of the RevBayes history before writing it to a file.
This will cause the RevBayes session to quit after evaluated the previous code, and is good practice
when running an mcmc in an RStudio terminal.}
}
\description{
Captures the current RevBayes history and saves it to an external file.
}
\details{
The user can then choose to run it in RevBayes directly with source,
or in an RStudio terminal via callRevFromTerminal().

Usage of this function can be combined with loadRev() to allow
for the continual use of the same RevBayes history between R sessions.
}
\examples{
\dontrun{
      saveRev("mySavedCode.rev", use_quit=FALSE)
      saveRev("mySavedCode.rev", use_wd=FALSE, use_quit=TRUE)
}


}
