% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRev.R
\name{saveRev}
\alias{saveRev}
\title{Save Revticulate history in a .Rev file for use in RevBayes}
\usage{
saveRev(filepath, use_wd = TRUE, use_quit = FALSE)
}
\arguments{
\item{filepath}{File to save .Revhistory to}
}
\description{
Save Revticulate history in a .Rev file for use in RevBayes
}
\examples{
\dontrun{
}


}
